Improve token auth header module
