+++ dmr/security/token/auth/header.py
@@ -6,8 +6,10 @@ from django.http import HttpRequest
 from typing_extensions import override
 
 from dmr.exceptions import NotAuthenticatedError
-from dmr.openapi.objects import Reference, SecurityRequirement, SecurityScheme
 from dmr.security.base import AsyncAuth, SyncAuth
+from dmr.security.token.auth.base import (
+    _BaseTokenAuth,  # noqa: WPS450  # pyright: ignore[reportPrivateUsage]
+)
 
 if TYPE_CHECKING:
     from django.contrib.auth.base_user import AbstractBaseUser
@@ -16,126 +18,6 @@ if TYPE_CHECKING:
     from dmr.endpoint import Endpoint
     from dmr.security.token.models import Token
     from dmr.serializer import BaseSerializer
-
-
-class _BaseTokenAuth:
-    """Base class for DB-backed opaque token authentication.
-
-    .. note::
-
-        By default every successful authentication performs an extra
-        ``UPDATE`` to persist ``last_used_at``.  Pass
-        ``update_last_used=False`` to skip this write for high-traffic
-        endpoints where tracking last-use is not required.
-    """
-
-    __slots__ = (
-        'cookie_name',
-        'header_name',
-        'prefix',
-        'query_param',
-        'security_scheme_name',
-        'update_last_used',
-    )
-
-    def __init__(
-        self,
-        *,
-        header_name: str = 'X-API-Token',
-        prefix: str = '',
-        query_param: str = 'token',
-        cookie_name: str = 'token',
-        security_scheme_name: str = 'token',
-        update_last_used: bool = True,
-    ) -> None:
-        """
-        Apply possible customizations.
-
-        - *header_name* — which header carries the raw token (default
-          ``X-API-Token``).  Use ``'Authorization'`` for RFC 7235-style
-          bearer auth (see *prefix* below).
-        - *prefix* — scheme prefix expected before the token in the header
-          value (default ``''``, i.e. the header value is used verbatim).
-          Set to ``'Token'`` or ``'Bearer'`` when *header_name* is
-          ``'Authorization'`` — e.g. ``prefix='Token'`` requires the client
-          to send ``Authorization: Token <raw-token>``.
-        - *query_param* — which query string parameter carries the raw token
-          (default ``token``).
-        - *cookie_name* — which cookie carries the raw token (default
-          ``token``).
-        - *security_scheme_name* — name used in OpenAPI security scheme map.
-
-        **Common configurations:**
-
-        .. code-block:: python
-
-            # Default — custom header, no prefix
-            TokenSyncAuth()  # X-API-Token: <token>
-
-            # DRF-compatible
-            TokenSyncAuth(header_name='Authorization', prefix='Token')
-
-            # Bearer style
-            TokenSyncAuth(header_name='Authorization', prefix='Bearer')
-        """
-        self.header_name = header_name
-        self.prefix = prefix
-        self.query_param = query_param
-        self.cookie_name = cookie_name
-        self.security_scheme_name = security_scheme_name
-        self.update_last_used = update_last_used
-
-    @property
-    def security_schemes(self) -> dict[str, SecurityScheme | Reference]:
-        """Provides a security schema definition."""
-        if self.header_name == 'Authorization':
-            # RFC 7235 reserves the Authorization header for the `http` scheme.
-            # Using `apiKey` + `name: Authorization` is invalid in OpenAPI 3.x.
-            return {
-                self.security_scheme_name: SecurityScheme(
-                    type='http',
-                    scheme='bearer',
-                    description='Opaque token authentication',
-                ),
-            }
-        return {
-            self.security_scheme_name: SecurityScheme(
-                type='apiKey',
-                name=self.header_name,
-                security_scheme_in='header',
-                description='Opaque token authentication',
-            ),
-        }
-
-    @property
-    def security_requirement(self) -> SecurityRequirement:
-        """Provides a security schema usage requirement."""
-        return {self.security_scheme_name: []}
-
-    def get_raw_token(self, request: HttpRequest) -> str | None:
-        """
-        Extract the raw token string from the request.
-
-        By default reads from the header specified by *header_name*, stripping
-        any *prefix*.  Override this method to read from a different source
-        (cookie, query string, request body, etc.).
-        """
-        header_value = request.headers.get(self.header_name)
-        if header_value is None:
-            return None
-        if self.prefix:
-            expected = f'{self.prefix} '
-            if not header_value.startswith(expected):
-                return None
-            return header_value[len(expected) :]
-        return header_value
-
-
-def _token_model() -> 'type[Token]':
-    """Deferred import so models are not loaded before app registry is ready."""
-    from dmr.security.token.models import Token  # noqa: PLC0415
-
-    return Token
 
 
 def _set_request_attrs(
@@ -179,6 +61,7 @@ class TokenSyncAuth(_BaseTokenAuth, SyncAuth):
     ) -> 'AbstractBaseUser':
         """Run all auth pipeline."""
         token = self.get_token(raw_token)
+        self.check_token(token)
         self.check_user(token.user)
         if self.update_last_used:
             self.mark_token_used(token)
@@ -187,7 +70,7 @@ class TokenSyncAuth(_BaseTokenAuth, SyncAuth):
 
     def get_token(self, raw_token: str) -> 'Token':
         """Look up and validate the token from the DB."""
-        model = _token_model()
+        model = self.token_model()
         token_hash = model.objects.hash_token(raw_token)
         try:
             token = model.objects.select_related('user').get(
@@ -195,9 +78,12 @@ class TokenSyncAuth(_BaseTokenAuth, SyncAuth):
             )
         except ObjectDoesNotExist:
             raise NotAuthenticatedError from None
+        return token
+
+    def check_token(self, token: 'Token') -> None:
+        """Raise NotAuthenticatedError if the token is not active."""
         if not token.is_active:
             raise NotAuthenticatedError
-        return token
 
     def check_user(self, user: 'AbstractBaseUser') -> None:
         """Raise NotAuthenticatedError if user account is not active."""
@@ -206,8 +92,12 @@ class TokenSyncAuth(_BaseTokenAuth, SyncAuth):
 
     def mark_token_used(self, token: 'Token') -> None:
         """Persist token usage timestamp after successful authentication."""
-        token.last_used_at = dt.datetime.now(dt.UTC)
-        token.save(update_fields=['last_used_at', 'updated_at'])
+        now = dt.datetime.now(dt.UTC)
+        self.token_model().objects.filter(pk=token.pk).update(
+            last_used_at=now,
+            updated_at=now,
+        )
+        token.last_used_at = now
 
     def set_request_attrs(
         self,
@@ -245,6 +135,7 @@ class TokenAsyncAuth(_BaseTokenAuth, AsyncAuth):
     ) -> 'AbstractBaseUser':
         """Run all auth pipeline."""
         token = await self.get_token(raw_token)
+        await self.check_token(token)
         await self.check_user(token.user)
         if self.update_last_used:
             await self.mark_token_used(token)
@@ -253,7 +144,7 @@ class TokenAsyncAuth(_BaseTokenAuth, AsyncAuth):
 
     async def get_token(self, raw_token: str) -> 'Token':
         """Look up and validate the token from the DB."""
-        model = _token_model()
+        model = self.token_model()
         token_hash = model.objects.hash_token(raw_token)
         try:
             token = await model.objects.select_related('user').aget(
@@ -261,9 +152,12 @@ class TokenAsyncAuth(_BaseTokenAuth, AsyncAuth):
             )
         except ObjectDoesNotExist:
             raise NotAuthenticatedError from None
+        return token
+
+    async def check_token(self, token: 'Token') -> None:
+        """Raise NotAuthenticatedError if the token is not active."""
         if not token.is_active:
             raise NotAuthenticatedError
-        return token
 
     async def check_user(self, user: 'AbstractBaseUser') -> None:
         """Raise NotAuthenticatedError if user account is not active."""
@@ -272,8 +166,12 @@ class TokenAsyncAuth(_BaseTokenAuth, AsyncAuth):
 
     async def mark_token_used(self, token: 'Token') -> None:
         """Persist token usage timestamp after successful authentication."""
-        token.last_used_at = dt.datetime.now(dt.UTC)
-        await token.asave(update_fields=['last_used_at', 'updated_at'])
+        now = dt.datetime.now(dt.UTC)
+        await self.token_model().objects.filter(pk=token.pk).aupdate(
+            last_used_at=now,
+            updated_at=now,
+        )
+        token.last_used_at = now
 
     async def set_request_attrs(
         self,
